pyproject.toml
[project]
name = "polyglot-grounded-qa"
version = "0.1.0"
description = "Notebook-first multilingual grounded QA with reusable core components"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "polyglot-grounded-qa contributors" }
]
dependencies = [
  "pydantic>=2.12.0",
  "pydantic-settings>=2.10.0",
  "pyyaml>=6.0.2",
  "polars>=1.33.0",
  "duckdb>=1.4.0",
  "pyarrow>=20.0.0",
  "numpy>=2.2.0",
  "pandas>=2.3.0",
]

[project.optional-dependencies]
notebooks = [
  "jupyterlab>=4.4.0",
  "jupytext>=1.17.0",
  "ipykernel>=6.30.0",
  "matplotlib>=3.10.0",
  "seaborn>=0.13.2",
]
retrieval = [
  "faiss-cpu>=1.12.0",
  "bm25s>=0.2.13",
  "sentence-transformers>=5.1.0",
  "lancedb>=0.25.0",
]
evaluation = [
  "ragas>=0.3.7",
  "scikit-learn>=1.7.0",
]
llm = [
  "openai>=2.0.0",
  "anthropic>=0.70.0",
  "google-genai>=1.30.0",
  "ollama>=0.5.3",
  "httpx>=0.28.0",
]
dev = [
  "pytest>=8.4.0",
  "pytest-cov>=6.2.0",
  "mypy>=1.17.0",
  "ruff>=0.12.0",
  "pre-commit>=4.3.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/polyglot_grounded_qa"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
disallow_untyped_defs = true
exclude = ["notebooks/"]

[tool.jupytext]
formats = "ipynb,py:percent"
